pyproject.toml
[project]
name = "exam-study-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "inngest>=0.5.15",
    "llama-index-core>=0.14.13",
    "llama-index-readers-file>=0.5.6",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.40.0",
]
